ai-agent/complex/pyproject.toml
[project]
name = "complex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=2.8.0",
    "langchain>=0.3.25",
    "langchain-openai>=0.3.24",
    "langgraph>=0.4.8",
    "pydentic>=0.0.1.dev3",
    "python-dotenv>=1.1.0",
]
